services/asr/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
redis==5.0.1
asyncpg==0.29.0
pydantic==2.9.2
numpy==1.26.4
openai-whisper==20231117
python-multipart==0.0.9
pydantic-settings==2.6.1
